refactor(src): align file structure with bulletproof-react
