--- src/app/components/PublishedDate.astro
@@ -1,25 +1,0 @@
----
-import { formatIsoString, formatYYMMDDString } from '../entry/date';
-
-interface Props {
-  publishedAt: number;
-  className?: string;
-}
-
-const { publishedAt, className } = Astro.props;
-
-const classNames = !className ? 'PublishedDate' : `PublishedDate ${className}}`;
-const dateTime = formatIsoString(publishedAt);
-const dateString = formatYYMMDDString(publishedAt);
----
-
-<style>
-  .PublishedDate {
-    font-size: var(--font-size-note);
-    line-height: var(--line-height-note);
-  }
-</style>
-
-<time class={classNames} datetime={dateTime}>
-  {dateString}
-</time>
