+++ src/components/EntryList.astro
@@ -1,7 +1,7 @@
 ---
-import { pathList } from '../../../constants/path_list';
-import { formatIsoString, formatYYMDString } from '../entry/date';
-import type { TinyCollectionEntry } from '../entry/tiny_collection_entry';
+import { pathList } from '../../constants/path_list';
+import { formatIsoString, formatYYMDString } from '../features/entry/date';
+import type { TinyCollectionEntry } from '../features/entry/tiny_collection_entry';
 
 interface Props {
   entries: TinyCollectionEntry[];
