+++ src/components/Logo.astro
@@ -1,6 +1,6 @@
 ---
-import { pathList } from '../../../constants/path_list';
-import { addWbr } from '../shared_logic/add_wbr';
+import { pathList } from '../../constants/path_list';
+import { addWbr } from '../utils/add_wbr';
 
 interface Props {
   title: string;
