+++ src/components/OpenGraphProtocol.astro
@@ -1,6 +1,6 @@
 ---
-import { FALLBACK_OG_IMAGE_URL, TWITTER_ACCOUNT_ID } from '../../../constants/site_data';
-import { retrieveTranslation } from '../locales/i18n';
+import { FALLBACK_OG_IMAGE_URL, TWITTER_ACCOUNT_ID } from '../../constants/site_data';
+import { retrieveTranslation } from '../features/locales/i18n';
 
 interface Props {
   title: string;
