+++ src/components/Switch.astro
