+++ src/features/entry/__fixtures__/i-entered-kua.md
