+++ src/features/entry/__mocks__/collection_entry.factory.ts
