+++ src/features/entry/get_sorted_entries.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from 'vitest';
 
-import { collectionEntryFactory } from '../__mocks__/collection_entry.factory';
-import { getSortedEntries } from '../get_sorted_entries';
-import type { TinyCollectionEntry } from '../tiny_collection_entry';
+import { collectionEntryFactory } from './__mocks__/collection_entry.factory';
+import { getSortedEntries } from './get_sorted_entries';
+import type { TinyCollectionEntry } from './tiny_collection_entry';
 
 function getCollectionEntry(publishedAt: Date): TinyCollectionEntry {
   const sharedProperty = {
