+++ src/features/likes/hooks/likes_buffer/internals/storage.ts
@@ -1,4 +1,4 @@
-import { getDOMStorage } from '../../../../../app/global_object/storage';
+import { getDOMStorage } from '../../../../../utils/global_object/storage';
 import { captureError } from '../../../../../utils/sentry';
 import type { RetryQueueItem } from './types';
 import { RETRY_QUEUE_KEY } from './types';
