+++ src/features/locales/__tests__/i18n.test.ts
