+++ src/features/locales/ja.ts
