+++ src/features/structured_data/home_structured_data.ts
@@ -1,5 +1,5 @@
 import { SITE_URL } from '../../../constants/site_data';
-import { retrieveTranslation } from '../locales/i18n';
+import { retrieveTranslation } from '../../features/locales/i18n';
 
 // https://developers.google.com/search/docs/appearance/site-names#website
 export interface HomeStructuredData {
