+++ src/pages/_layouts/EntryShell.astro
@@ -1,7 +1,7 @@
 ---
 import { pathList } from '../../../constants/path_list';
 import { OG_IMAGE_PATH, SITE_URL } from '../../../constants/site_data';
-import OpenGraphProtocol from '../../app/components/OpenGraphProtocol.astro';
+import OpenGraphProtocol from '../../components/OpenGraphProtocol.astro';
 import { SlotNames } from '../_shared/slot_names';
 import BaseShell from './BaseShell.astro';
