+++ src/pages/index.astro
@@ -2,12 +2,12 @@
 import { getCollection } from 'astro:content';
 
 import { pathList } from '../../constants/path_list';
-import EntryList from '../app/components/EntryList.astro';
-import Logo from '../app/components/Logo.astro';
-import { addExcerptToEntries } from '../app/entry/add_excerpt_to_entries';
-import { getSortedEntries } from '../app/entry/get_sorted_entries';
-import { retrieveTranslation } from '../app/locales/i18n';
-import { createHomeStructuredData } from '../app/structured_data/home_structured_data';
+import EntryList from '../components/EntryList.astro';
+import Logo from '../components/Logo.astro';
+import { addExcerptToEntries } from '../features/entry/add_excerpt_to_entries';
+import { getSortedEntries } from '../features/entry/get_sorted_entries';
+import { retrieveTranslation } from '../features/locales/i18n';
+import { createHomeStructuredData } from '../features/structured_data/home_structured_data';
 import BaseShell from './_layouts/BaseShell.astro';
 import { SlotNames } from './_shared/slot_names';
