+++ src/pages/policy/index.astro
@@ -2,7 +2,7 @@
 import '../_shared/entry_contents.css';
 
 import { pathList } from '../../../constants/path_list';
-import { retrieveTranslation } from '../../app/locales/i18n';
+import { retrieveTranslation } from '../../features/locales/i18n';
 import BaseShell from '../_layouts/BaseShell.astro';
 import AnalyticsOptoutSwitch from './_components/AnalyticsOptoutSwitch.astro';
