+++ src/pages/tags/[tag].astro
@@ -2,10 +2,10 @@
 import { getCollection } from 'astro:content';
 
 import { pathList } from '../../../constants/path_list';
-import EntryList from '../../app/components/EntryList.astro';
-import { addExcerptToEntries } from '../../app/entry/add_excerpt_to_entries';
-import { getSortedEntries } from '../../app/entry/get_sorted_entries';
-import { retrieveTranslation } from '../../app/locales/i18n';
+import EntryList from '../../components/EntryList.astro';
+import { addExcerptToEntries } from '../../features/entry/add_excerpt_to_entries';
+import { getSortedEntries } from '../../features/entry/get_sorted_entries';
+import { retrieveTranslation } from '../../features/locales/i18n';
 import BaseShell from '../_layouts/BaseShell.astro';
 
 export async function getStaticPaths() {
