+++ src/utils/global_object/dom_storage_like.ts
