+++ src/utils/global_object/storage.ts
