+++ src/utils/global_object/window.ts
